架构层面的重大改动：   1. 从 Feign 同步调用改为 RabbitMQ 事件驱动 — 删除了各模块的 FeignRequestInterceptor、InterceptorConfig 和 Feign 接口，新增了 RabbitMQConfig 及 event/listener 包，服务间通信从 同步改为异步消息。   2. 新增 ai-server、gateway-server、search-server 三个独立模块 — 引入了 API 网关和新服务。   3. 将 AI 审核从 common-server 独立到 ai-server — 删除了 AiAuditController、AiAuditService 等，移到了新模块。
  功能层面的重大改动：
  4. 新增 novel-ai-agent 模块 — 提供 AI 写作辅助（润色、续写、知识提取等）。
  5. 新增 InternalController 类 — 各模块新增内部接口用于服务间调用。
  6. 前端新增 AI 写作、排行榜、阅读进度、搜索等功能页面和 API。
  7. 从 common 模块移除了 RoleInterceptor，新增了 common/feign/ 和 common/filter/ 基础设施。
  8. 新增 KnowledgeItem、VisitorReadingProgress 实体及相关 DTO/VO。

  总的来说，这是一次从 Feign 同步微服务 → RabbitMQ 异步事件驱动 + API 网关 的大架构升级，同时加入了 AI 写作助手能力。

+++ author-server/src/main/java/com/wang/manage/controller/AuthorController.java
@@ -35,20 +35,20 @@ public class AuthorController {
 
     @PostMapping("logout")
     @ApiOperation("作者退出登入")
-    public Result logout(Integer id){
+    public Result logout(Long id){
         log.info("作者退出登入,id= {}", id);
         return authorService.logout(id);
     }
 
     @GetMapping("info/{id}")
     @ApiOperation("获取作者信息")
-    public Result getAuthorInfo(@PathVariable Integer id){
+    public Result getAuthorInfo(@PathVariable Long id){
         log.info("获取作者信息请求：ID={}", id);
         return authorService.getAuthorInfo(id);
     }
     @GetMapping("getNameAndAvatar/{id}")
     @ApiOperation("获取作者公开信息（访客端）")
-    public Result getNameAndAvatar(@PathVariable Integer id){
+    public Result getNameAndAvatar(@PathVariable Long id){
         log.info("获取作者公开信息请求：ID={}", id);
         return authorService.getNameAndAvatar(id);
     }
@@ -64,7 +64,7 @@ public class AuthorController {
 
     @DeleteMapping("delete/{id}")
     @ApiOperation("作者注销（逻辑删除）")
-    public Result deleteAuthor(@PathVariable Integer id){
+    public Result deleteAuthor(@PathVariable Long id){
         log.info("作者注销请求：ID={}", id);
         return authorService.deleteAuthor(id);
     }
@@ -81,7 +81,7 @@ public class AuthorController {
     @PostMapping("/updatePassword")
     @ApiOperation("修改作者密码")
     public Result updatePassword(
-            @ApiParam("作者ID") @RequestParam Integer id,
+            @ApiParam("作者ID") @RequestParam Long id,
             @ApiParam("旧密码") @RequestParam String oldPassword,
             @ApiParam("新密码") @RequestParam String newPassword) {
         log.info("修改作者密码请求：ID={}", id);
