+++ common-server/src/main/java/com/wang/commonserver/service/impl/FileServiceImpl.java
@@ -34,7 +34,7 @@ public class FileServiceImpl implements FileService {
     }
 
     @Override
-    public Result uploadFile(MultipartFile file, Integer code, Integer novelId, String oldFileUrl) {
+    public Result uploadFile(MultipartFile file, Integer code, Long novelId, String oldFileUrl) {
         String typeName = FileUploadTypeEnum.getMessageByCode(code).name();
         if (!StringUtils.hasText(typeName)) {
             log.info("无效的类型:code={}",code);
