+++ visitor-server/src/main/java/com/wang/visitor/controller/VisitorFollowController.java
@@ -36,8 +36,8 @@ public class VisitorFollowController {
     @DeleteMapping("/remove/{authorId}")
     @ApiOperation("取消关注")
     public Result unfollow(
-            @PathVariable @ApiParam("作者ID") Integer authorId,
-            @RequestParam @ApiParam("访客ID") Integer visitorId) {
+            @PathVariable @ApiParam("作者ID") Long authorId,
+            @RequestParam @ApiParam("访客ID") Long visitorId) {
         log.info("取消关注请求：visitorId={}, authorId={}", visitorId, authorId);
         return visitorFollowService.unfollow(visitorId, authorId);
     }
@@ -45,8 +45,8 @@ public class VisitorFollowController {
     @GetMapping("/check/{authorId}")
     @ApiOperation("检查是否已关注")
     public Result checkFollow(
-            @PathVariable @ApiParam("作者ID") Integer authorId,
-            @RequestParam @ApiParam("访客ID") Integer visitorId) {
+            @PathVariable @ApiParam("作者ID") Long authorId,
+            @RequestParam @ApiParam("访客ID") Long visitorId) {
         log.info("检查是否关注请求：visitorId={}, authorId={}", visitorId, authorId);
         return visitorFollowService.checkFollow(visitorId, authorId);
     }
@@ -54,7 +54,7 @@ public class VisitorFollowController {
     @GetMapping("/list")
     @ApiOperation("获取我的关注列表")
     public Result getMyFollows(
-            @RequestParam @ApiParam("访客ID") Integer visitorId,
+            @RequestParam @ApiParam("访客ID") Long visitorId,
             @RequestParam(defaultValue = "1") @ApiParam("页码") Integer pageNum,
             @RequestParam(defaultValue = "10") @ApiParam("每页数量") Integer pageSize) {
         log.info("获取我的关注列表请求：visitorId={}", visitorId);
@@ -64,7 +64,7 @@ public class VisitorFollowController {
     @GetMapping("/followers/{authorId}")
     @ApiOperation("获取作者的粉丝列表")
     public Result getFollowers(
-            @PathVariable @ApiParam("作者ID") Integer authorId,
+            @PathVariable @ApiParam("作者ID") Long authorId,
             @RequestParam(defaultValue = "1") @ApiParam("页码") Integer pageNum,
             @RequestParam(defaultValue = "10") @ApiParam("每页数量") Integer pageSize) {
         log.info("获取作者粉丝列表请求：authorId={}", authorId);
@@ -73,14 +73,14 @@ public class VisitorFollowController {
 
     @GetMapping("/count")
     @ApiOperation("获取我的关注数量")
-    public Result getMyFollowCount(@RequestParam @ApiParam("访客ID") Integer visitorId) {
+    public Result getMyFollowCount(@RequestParam @ApiParam("访客ID") Long visitorId) {
         log.info("获取我的关注数量请求：visitorId={}", visitorId);
         return visitorFollowService.getMyFollowCount(visitorId);
     }
 
     @GetMapping("/followerCount/{authorId}")
     @ApiOperation("获取作者的粉丝数量")
-    public Result getFollowerCount(@PathVariable @ApiParam("作者ID") Integer authorId) {
+    public Result getFollowerCount(@PathVariable @ApiParam("作者ID") Long authorId) {
         log.info("获取作者粉丝数量请求：authorId={}", authorId);
         return visitorFollowService.getFollowerCount(authorId);
     }
